+++ lib/basket/testing.rb
@@ -1,0 +1,4 @@
+module Basket
+  module Testing
+  end
+end
