+++ spec/fixtures/fireworks.rb
@@ -1,0 +1,8 @@
+class DummyFireworksBasket
+  include Basket::Batcher
+  basket_options size: 1
+
+  def perform
+    raise "Boom"
+  end
+end
